+++ delivery/services/file_system_service.py
@@ -1,0 +1,45 @@
+
+import os
+
+import logging
+
+log = logging.getLogger(__name__)
+
+
+class FileSystemService(object):
+    """
+    File system service, used for accessing the file system in a way that can
+    easily be mocked out in testing.
+    """
+
+    @staticmethod
+    def _list_directories(base_path):
+        """
+        List all directories
+        :param base_path: base path to list directories in.
+        :return: a generator of paths to directories
+        """
+
+        log.debug("Listing dirs in: {}".format(os.path.abspath(base_path)))
+        for my_dir in os.listdir(base_path):
+            dir_abs_path = os.path.abspath(os.path.join(base_path, my_dir))
+
+            if os.path.isdir(dir_abs_path):
+                log.debug("Found dir: {}".format(dir_abs_path))
+                yield dir_abs_path
+
+    def find_project_directories(self, projects_base_dir):
+        """
+        Find project directories
+        :param projects_base_dir: directory to list
+        :return: a generator of paths to project directories
+        """
+        return self._list_directories(projects_base_dir)
+
+    def find_runfolder_directories(self, base_path):
+        """
+        Find runfolder directories
+        :param base_path: directory to list
+        :return: a generator or paths to runfolder directories
+        """
+        return self._list_directories(base_path)
